Use the assembly resolver from the module definition
Use the IAssemblyResolver provided by the module definition itself instead of the default resolver when reading the Costura template with Cecil so that assembly references are properly resolved for .NET Framework assemblies, even when running from a .NET Core runtime (which is exactly what is happening when running the `dotnet msbuild` command).

When using the default assembly resolver, bad assembly references would be added to a weaved .NET Framework assembly, for example `System.Private.CoreLib` would be added as an assembly reference, making the weaved executable crash at startup.

Fixes #694