Use a special assembly resolver to support .NET Framework 4.7 and lower
The change introduced in #696 only worked for apps using Costura on a .NET Core target framework or .NET Framework 4.7.1 and later. When using Costura for a project targeting .NET Framework 4.7.0 or lower, the following exception would be thrown:
```
System.ArgumentNullException
   at Mono.Cecil.Mixin.CheckType(Object type)
   at Mono.Cecil.ModuleDefinition.ImportReference(TypeReference type, IGenericParameterProvider context)
   at Mono.Cecil.ModuleDefinition.ImportReference(TypeReference type)
   at ModuleWeaver.Resolve(TypeReference baseType)
   at ModuleWeaver.ImportAssemblyLoader(Boolean createTemporaryAssemblies)
   at ModuleWeaver.Execute()
   at InnerWeaver.ExecuteWeavers()
```

It is not clear from the exception, but it occurs because the assembly resolver can't resolve the `netstandard` assembly reference. To properly resolve the `netstandard` reference on .NET Framework versions where it's not available, we embed the one from the .NET Framework 4.8 assembly references.

Fixes #697